jobs/resume-parser/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.18
google-cloud-documentai>=2.29
